pyproject.toml
[project]
name = "locust-tester"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=40.1.0",
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "locust>=2.43.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
